% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmypath.R
\name{findmypath}
\alias{findmypath}
\title{Find path to graphiT package}
\usage{
findmypath(dir, file)
}
\description{
This function is used internally to find package's path.
}
\keyword{graphiT,}
\keyword{graphic,}
\keyword{toolkat}

