% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphiT.R
\name{graphiT}
\alias{graphiT}
\title{Launch the Graphic ToolKat}
\usage{
graphiT()
}
\description{
This function allows you to launch the Graphic ToolKat. This is a Shiny App designed to help users produce ggplot graphics.
}
\examples{
graphiT()
}
\keyword{bwegek}

