% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmypath.R
\name{findmypath}
\alias{findmypath}
\title{BWEGEGEK!}
\usage{
findmypath(dir, file)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

