% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphiT.R
\name{graphiT}
\alias{graphiT}
\title{Launch the Graphic ToolKat}
\usage{
graphiT()
}
\description{
This function allows you to launch the Graphic ToolKat interface. This is a Shiny App designed to help users produce graphics.
For more information see package's vignette.
}
\examples{
graphiT()
}
\keyword{graphiT,}
\keyword{graphic,}
\keyword{toolkat}

